rename: 두 페이지에서 사용하는 공통 훅 디렉토리 위치 변경
